6. practice

+++ 2020-21-2/webeng1/mytracks/app/Http/Controllers/ProjectController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\ProjectFormRequest;
 use App\Models\Project;
 use GuzzleHttp\Handler\Proxy;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
@@ -14,12 +15,13 @@ class ProjectController extends Controller
         $projects = Project::all();
         // dd($projects);
         return view('projects.list', [
-            'projects'  => $projects,
+            'projects'  => Auth::user()->projects,
         ]);
     }
 
     public function show(Project $project) // list one, display page
     {
+        $this->authorize('access', $project);
         // $project = Project::find($id);
         return view('projects.show', [
             'project'  => $project,
@@ -34,12 +36,14 @@ class ProjectController extends Controller
     public function store(ProjectFormRequest $request)
     {
         $validated_data = $request->validated();
+        $validated_data['user_id'] = Auth::id();
         Project::create($validated_data);
         return redirect()->route('projects.index');
     }
 
     public function edit(Project $project)
     {
+        $this->authorize('access', $project);
         // dump($id);
         // $project = Project::find($id);
         // dd($project);
@@ -50,6 +54,7 @@ class ProjectController extends Controller
 
     public function update(Project $project, ProjectFormRequest $request)
     {
+        $this->authorize('access', $project);
         $validated_data = $request->validated();
         // $project = Project::find($id);
         $project->update($validated_data);
@@ -59,6 +64,7 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        $this->authorize('access', $project);
         // Project::destroy($id);
         $project->delete();
         return redirect()->route('projects.index');
