+++ 2020-21-2/webeng1/mytracks/app/Http/Controllers/TrackController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Filter;
 use App\Models\Project;
 use App\Models\Track;
 use Illuminate\Http\Request;
@@ -15,8 +16,11 @@ class TrackController extends Controller
      */
     public function create(Project $project)
     {
+        $this->authorize('access', $project);
+        $filters = Filter::all();
         return view('tracks.create', [
             'project_id' => $project->id,
+            'filters' => $filters,
         ]);
     }
 
@@ -28,11 +32,16 @@ class TrackController extends Controller
      */
     public function store(Project $project, Request $request)
     {
+        $this->authorize('access', $project);
         $data = $request->validate([
             'name'  => 'required',
             'color' => 'required|regex:/^#[0-9a-z]{6}$/',
+            'filters' => 'nullable|array'
         ]);
-        $project->tracks()->create($data);
+        $track = $project->tracks()->create($data);
+        if (isset($data['filters'])) {
+            $track->filters()->attach($data['filters']);
+        }
         return redirect()->route('projects.show', ['project' => $project->id]);
     }
 
@@ -44,8 +53,12 @@ class TrackController extends Controller
      */
     public function edit(Track $track)
     {
+        $this->authorize('access', $track);
+        $filters = Filter::all();
+        $track->load('filters');
         return view('tracks.edit', [
             'track' => $track,
+            'filters' => $filters,
         ]);
     }
 
@@ -58,11 +71,14 @@ class TrackController extends Controller
      */
     public function update(Request $request, Track $track)
     {
+        $this->authorize('access', $track);
         $data = $request->validate([
             'name'  => 'required',
             'color' => 'required|regex:/^#[0-9a-z]{6}$/',
+            'filters' => 'nullable|array',
         ]);
         $track->update($data);
+        $track->filters()->sync($data['filters'] ?? []);
         return redirect()->route('projects.show', ['project' => $track->project_id]);
     }
 
@@ -74,6 +90,7 @@ class TrackController extends Controller
      */
     public function destroy(Track $track)
     {
+        $this->authorize('access', $track);
         $track->delete();
         return redirect()->route('projects.show', ['project' => $track->project_id]);
     }
